corp_num 타입 변경 및 프론트에 맞게 백엔드 일부 수정

+++ src/models/rdbms/Corporation.ts
@@ -110,7 +110,7 @@ export class Corporation
                     allowNull: true,
                 },
                 corp_num: {
-                    type: DataTypes.INTEGER.UNSIGNED,
+                    type: DataTypes.BIGINT.UNSIGNED,
                     allowNull: false,
                 },
                 biz_num: {
